src/requirements.txt
uvicorn==0.28.0
pydantic==2.5.3
aiogram==3.4.1
fastapi==0.110.0
pydantic-settings==2.2.1
tortoise-orm==0.20.0
asyncpg==0.27.0
aerich==0.7.1
asyncio-redis==0.16.0
